refactoring code a bit
